-- CRM-10372 GroupBy code cleanup and orderBy improvements

+++ CRM/Campaign/BAO/Survey.php
@@ -399,8 +399,9 @@ SELECT  survey.id    as id,
       return NULL;
     }
     $reportId = CRM_Campaign_BAO_Survey::getReportID($id);
-    CRM_Report_BAO_Instance::delete($reportId);
-    
+    if($reportId){
+      CRM_Report_BAO_Instance::delete($reportId);
+    }
     $dao = new CRM_Campaign_DAO_Survey();
     $dao->id = $id;
     return $dao->delete();
